fix: Audio preview got replayed upon changing Queue pagination
